Update for win32 build: remove unused variables and unpublished method.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@4349 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/snmplib/read_config.c
@@ -818,10 +818,10 @@ char *read_config_save_octet_string(char *saveto, u_char *str, size_t len) {
   char *cp;
 
   /* is everything easily printable */
-  for(i=0, cp=str; i < len && cp &&
+  for(i=0, cp=str; i < (int)len && cp &&
           (isalpha(*cp) || isdigit(*cp) || *cp == ' '); cp++, i++);
 
-  if (len != 0 && i == len) {
+  if (len != 0 && i == (int)len) {
       sprintf(saveto, "\"%s\"", str);
       saveto += strlen(saveto);
   } else {
